Create CX JSON text representing the network

Create a CX JSON text that represent the network which includes information about nodes and edges present in the network.

createCytoscapeCXJSON(gostResults, gostObject, title)

Arguments

gostResults

a data.frame containing the terms retained for the creation of the network.

gostObject

a list created by gprofiler2 that contains the results from an enrichment analysis.

title

a character string representing the name assigned to the network.

Value

character string that represent the network in a CX JSON format.

Author

Astrid Deschênes

Examples

## Loading dataset containing result from an enrichment analysis done with
## gprofiler2
data(demoGOST)

## Only retained the WikiPathways results
results <- demoGOST$result[demoGOST$result$source == "WP", ]

jsonFormat <- enrichViewNet:::createCytoscapeCXJSON(
                gostResults = results, gostObject = demoGOST, 
                title = "WikiPathways")