From functional enrichment results to biological networks

The enrichViewNet package enables the visualization of enrichment results, in a format corresponding to the one generated by gprofiler2 (https://cran.r-project.org/web/packages/gprofiler2/index.html) under the form of Cytoscape network (https://cytoscape.org/).

Details

In those networks, both gene datasets (GO terms/pathways/protein complexes) and genes are represented as nodes. A edge connect a gene to its datasets. In the current version, only genes present in at least one gene dataset are retained.

See also

Author

Astrid Deschênes, Pascal Belleau, Robert L Faure, Maria J Fernandes, David A Tuveson

Maintainer: Astrid Deschênes <adeschen@hotmail.com>