Using functional enrichment results from gprofiler2 to create a Cytoscape network

User selected enrichment terms are used to create a Cytoscape network where the selected terms and the genes that where part of the enrichment analysis are all represented as nodes. Edges are linking the genes to their terms. The selection of the term can by specifying the source of the terms (GO:MF, REAC, TF, etc...) or by listing the selected term IDs. The network is only generated when there is at least on significant term to graph. When the enrichment analysis contains more than one query, only one query can be selected to generate the network.

createNetwork(
  gostObject,
  source = c("TERM_ID", "GO:MF", "GO:CC", "GO:BP", "KEGG", "REAC", "TF", "MIRNA", "HPA",
    "CORUM", "HP", "WP"),
  termIDs = NULL,
  removeRoot = TRUE,
  query = NULL,
  title = "gprofiler network",
  collection = "enrichment results",
  fileName = "gprofilerNetwork.cx"
)

Arguments

gostObject

a list created by gprofiler2 that contains the results from an enrichment analysis.

source

a character string representing the selected source that will be used to generate the network. To hand-pick the terms to be used, "TERM_ID" should be used and the list of selected term IDs should be passed through the termIDs parameter. The possible sources are "GO:BP" for Gene Ontology Biological Process, "GO:CC" for Gene Ontology Cellular Component, "GO:MF" for Gene Ontology Molecular Function, "KEGG" for Kegg, "REAC" for Reactome, "TF" for TRANSFAC, "MIRNA" for miRTarBase, "CORUM" for CORUM database, "HP" for Human phenotype ontology and "WP" for WikiPathways. Default: "TERM_ID".

termIDs

a vector of character strings that contains the term IDS retained for the creation of the network. Default: NULL.

removeRoot

a logical that specified if the root terms of the selected source should be removed (when present). Default: TRUE.

query

a character string that specified the retained query to generate the network. When NULL, the query present in the result is retained; NULL cannot be used when more than one query is present. Default: NULL.

title

a character string representing the name assigned to the network. Default: "gprofiler network".

collection

a character string representing the collection name assigned to the network. Default: "enrichment results".

fileName

a character string representing the name of the CX JSON file that is created when Cytoscape is not running. The name must have a '.cx' extension. Default: "gprofilerNetwork_01.cx".

Value

TRUE

Author

Astrid Deschênes

Examples

## Loading dataset containing result from an enrichment analysis done with
## gprofiler2
data(demoGOST)

## Some of the enrichment results present in the dataset
head(demoGOST$result)
#>     query significant      p_value term_size query_size intersection_size
#> 1 query_1        TRUE 2.777564e-02         1         15                 1
#> 2 query_1        TRUE 6.010214e-05      4700         15                13
#> 3 query_1        TRUE 6.836641e-05      2812         15                11
#> 4 query_1        TRUE 9.082557e-05      4857         15                13
#> 5 query_1        TRUE 1.466212e-04      5046         15                13
#> 6 query_1        TRUE 1.484313e-04      2264         15                10
#>    precision      recall    term_id source
#> 1 0.06666667 1.000000000 CORUM:7486  CORUM
#> 2 0.86666667 0.002765957 GO:0051171  GO:BP
#> 3 0.73333333 0.003911807 GO:0010604  GO:BP
#> 4 0.86666667 0.002676549 GO:0080090  GO:BP
#> 5 0.86666667 0.002576298 GO:0031323  GO:BP
#> 6 0.66666667 0.004416961 GO:0035556  GO:BP
#>                                                term_name effective_domain_size
#> 1                                              HAS2-HAS3                 24109
#> 2      regulation of nitrogen compound metabolic process                 24109
#> 3 positive regulation of macromolecule metabolic process                 24109
#> 4                regulation of primary metabolic process                 24109
#> 5               regulation of cellular metabolic process                 24109
#> 6                      intracellular signal transduction                 24109
#>   source_order      parents
#> 1         2858 CORUM:00....
#> 2        15706 GO:00068....
#> 3         4771 GO:00098....
#> 4        20968 GO:00192....
#> 5         8293 GO:00192....
#> 6        10508   GO:0007165

## Create network for Gene Ontology - Molecular Function related results
## in Cytoscape (when the application is opened)
## Otherwise, create a CX file in the temporary directory
## The file can be opened in Cytoscape
createNetwork(gostObject=demoGOST, source="GO:MF", removeRoot=FALSE,
    title="GO Molecular Function Graph", 
    fileName=file.path(tempdir(), "GO_MF_demo.cx"))
#> Oh no! I can't find Cytoscape. RCy3 can not continue!
#> Please check that Cytoscape is running, CyREST is installed and your base.url parameter is correct.
#> Unable to connect to Cytoscape. 
#> CX JSON file will be created.
#> CX JSON file "/tmp/RtmpRIP5q9/GO_MF_demo.cx" has been created.
#> [1] TRUE